third_party/blink/web_tests/fast/dom/icon-url-change-expected.txt

PASS successfullyParsed is true

TEST COMPLETE

Original iconURL is: http://test.com/oldfavicon.ico
Setting new icon URL to: http://test.com/newfavicon.ico
New iconURL is: http://test.com/newfavicon.ico
Setting icon URL back to: http://test.com/oldfavicon.ico
Original iconURL is still: http://test.com/oldfavicon.ico
PASS URL list matches expected
